move _client_projection() from mongo to DataLayer
_client_projection() is generic. There is no reason to keep it in
the mongodb drive. By moving it on the DataLayer class we will be
able to drop some (oudated) duplicated code from Eve-SQLAlchemy:

https://github.com/RedTurtle/eve-sqlalchemy/blob/0.4.0a2/eve_sqlalchemy/__init__.py#L325-L340
